cloud_run_publisher/requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-cloud-pubsub==2.31.1
